refactor: Long 타입의 ID를 받게끔 변경

+++ app/src/main/java/woowacourse/shopping/data/DefaultCartRepository.kt
@@ -10,5 +10,5 @@ class DefaultCartRepository(
 
     override suspend fun getAllCartProducts(): List<Product> = dao.getAll().map(CartProductEntity::toDomain)
 
-    override suspend fun deleteCartProduct(id: Int) = dao.delete(id.toLong())
+    override suspend fun deleteCartProduct(id: Long) = dao.delete(id.toLong())
 }
